structure the app to be a salesforce dx project
