--- ServiceDocumentSampleComponents/demoWorkPlan/__tests__/demoWorkPlan.test.js
@@ -1,25 +1,0 @@
-import { createElement } from 'lwc';
-import DemoWorkPlan from 'c/demoWorkPlan';
-
-describe('c-demo-work-plan', () => {
-    afterEach(() => {
-        // The jsdom instance is shared across test cases in a single file so reset the DOM
-        while (document.body.firstChild) {
-            document.body.removeChild(document.body.firstChild);
-        }
-    });
-
-    it('TODO: test case generated by CLI command, please fill in test logic', () => {
-        // Arrange
-        const element = createElement('c-demo-work-plan', {
-            is: DemoWorkPlan
-        });
-
-        // Act
-        document.body.appendChild(element);
-
-        // Assert
-        // const div = element.shadowRoot.querySelector('div');
-        expect(1).toBe(1);
-    });
-});
