--- ServiceDocumentSampleComponents/demoWorkPlan/demoWorkPlan.css
@@ -1,3 +1,0 @@
-:host h4, h3 {
-    color: #009ad7;
-}
