--- ServiceDocumentSampleComponents/demoWorkStepPresentation/demoWorkStepPresentation.css
@@ -1,3 +1,0 @@
-:host {
-    display: contents;
-  }
