--- ServiceDocumentSampleComponents/demoWorkStepPresentation/demoWorkStepPresentation.html
@@ -1,7 +1,0 @@
-<template>
-    <tr>
-        <td class="slds-cell-wrap">{workStepRecord.Name.value}</td>
-        <td class="slds-cell-wrap">{workStepRecord.Status.value}</td>
-        <td class="slds-cell-wrap">{workStepRecord.Description.value}</td>
-    </tr>    
-</template>
