--- ServiceDocumentSampleComponents/jsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "experimentalDecorators": true
-  },
-  "include": ["**/*", "../../../../.sfdx/typings/lwc/**/*.d.ts"],
-  "paths": {
-    "c/*": ["*"]
-  },
-  "typeAcquisition": {
-    "include": ["jest"]
-  },
-  "experimentalDecorators": true
-}
