+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/.husky/pre-commit
@@ -1,0 +1,4 @@
+#!/bin/sh
+. "$(dirname "$0")/_/husky.sh"
+
+npm run precommit
