+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/.prettierignore
@@ -1,0 +1,10 @@
+# List files or directories below to ignore them when running prettier
+# More information: https://prettier.io/docs/en/ignore.html
+#
+
+**/staticresources/**
+.localdevserver
+.sfdx
+.vscode
+
+coverage/
