+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/.vscode/extensions.json
@@ -1,0 +1,9 @@
+{
+  "recommendations": [
+    "salesforce.salesforcedx-vscode",
+    "redhat.vscode-xml",
+    "dbaeumer.vscode-eslint",
+    "esbenp.prettier-vscode",
+    "financialforce.lana"
+  ]
+}
