+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/config/project-scratch-def.json
@@ -1,0 +1,13 @@
+{
+  "orgName": "giaoge.zheng company",
+  "edition": "Developer",
+  "features": ["EnableSetPasswordInApi"],
+  "settings": {
+    "lightningExperienceSettings": {
+      "enableS1DesktopEnabled": true
+    },
+    "mobileSettings": {
+      "enableS1EncryptedStoragePref2": false
+    }
+  }
+}
