+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/force-app/main/default/aura/.eslintrc.json
@@ -1,0 +1,8 @@
+{
+  "plugins": ["@salesforce/eslint-plugin-aura"],
+  "extends": ["plugin:@salesforce/eslint-plugin-aura/recommended"],
+  "rules": {
+    "vars-on-top": "off",
+    "no-unused-expressions": "off"
+  }
+}
