+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/force-app/main/default/lwc/.eslintrc.json
@@ -1,10 +1,13 @@
 {
-  "extends": ["@salesforce/eslint-config-lwc/recommended", "prettier"],
+  "extends": ["@salesforce/eslint-config-lwc/recommended"],
   "overrides": [
     {
       "files": ["*.test.js"],
       "rules": {
         "@lwc/lwc/no-unexpected-wire-adapter-usages": "off"
+      },
+      "env": {
+        "node": true
       }
     }
   ]
