+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/force-app/main/default/lwc/bodyText/bodyText.js
@@ -1,5 +1,5 @@
 import { LightningElement, api } from "lwc";
 
 export default class BodyText extends LightningElement {
-    @api text;
+  @api text;
 }
