+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/force-app/main/default/lwc/demoWorkPlan/demoWorkPlan.css
@@ -1,0 +1,4 @@
+:host h4,
+h3 {
+  color: #009ad7;
+}
