+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/force-app/main/default/lwc/demoWorkStepPresentation/demoWorkStepPresentation.css
@@ -1,0 +1,3 @@
+:host {
+  display: contents;
+}
