+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/force-app/main/default/lwc/mediumHeaderText/mediumHeaderText.html
