+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/force-app/main/default/lwc/mediumHeaderText/mediumHeaderText.js-meta.xml
