+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/force-app/main/default/lwc/staticImage/staticImage.html
