+++ ServiceDocumentSampleComponents/serviceDocumentSampleComponents/jest.config.js
@@ -1,0 +1,6 @@
+const { jestConfig } = require("@salesforce/sfdx-lwc-jest/config");
+
+module.exports = {
+  ...jestConfig,
+  modulePathIgnorePatterns: ["<rootDir>/.localdevserver"]
+};
